requirements.txt
tensorflow>=2.12.0
numpy>=1.24.0
Pillow>=9.5.0